Crate rugra

Rugra is a ultra minimal graphics engine for Rust.

Modules

circle
Circles

font
Fonts

key
Keys

label
Labels

prelude
The prelude which reexports all the basic functionality Rugra contains.

rect
Rectangles

sound
Sound

sprite
Sprites

texture
Textures

window
Windows